[Fix] : /search/result 페이지 컴포넌트에서 useSearchParams() 훅을 사용하고 있어 이를 Suspense 컴포넌트로 래핑
1. dev 환경에서는 Suspense가 없어도 잘 돌아가는데, build 시에는 Suspense가 감싸줘야함
2. SearchResult 컴포넌트를 동일 경로에 만들어준 다음, page.tsx에서는 이를 import하고 Suspense로 감싸줌
